Fix -Wfloat-conversion and -Wdouble-promotion warnings
Use explicit casts when converting between double and float types.
Expression::complex_type is std::complex<float>, so multiplying by
double values requires explicit conversion.

Also fixes -Wshadow warnings in HubbardModel constructors by renaming
parameters to avoid shadowing member variables.

+++ src/algebra/fourier_transform.h
@@ -65,7 +65,9 @@ inline Expression fourier_transform_operator(Operator op, const Index& index) {
     coefficient = std::exp(coefficient) * normalization;
 
     Operator transformed_op(op.type(), op.spin(), k);
-    result += Term(Term::complex_type(coefficient.real(), coefficient.imag()), {transformed_op});
+    result += Term(Term::complex_type(static_cast<float>(coefficient.real()),
+                                      static_cast<float>(coefficient.imag())),
+                   {transformed_op});
   }
 
   return result;
